feat: 계층 구조 변경에 따른 도메인 계층 변경 - entity: 기존 entity 패키지 -> domain 하위 - transportcompany: aggregate 패키지 생성 (deliverydriver -> 해당 패키지로 이동) - service 분화 :   - 기존 service 클래스: 인터페이스와 구현체로 구성   - service의 기능에 따른 특화 클래스 생성: reader, persister -> 인터페이스로 도메인 계층에 존재 -> 세부 구현체들은 infrastruce으로 밀집
