+++ src/main/java/timf/voc/task/domain/claim/ClaimPersister.java
@@ -1,0 +1,5 @@
+package timf.voc.task.domain.claim;
+
+public interface ClaimPersister {
+	void save(Claim claim);
+}
