+++ src/main/java/timf/voc/task/domain/claim/ClaimReader.java
@@ -1,0 +1,10 @@
+package timf.voc.task.domain.claim;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface ClaimReader {
+	List<Claim> get();
+
+	Optional<Claim> get(Long claimId);
+}
