+++ src/main/java/timf/voc/task/domain/clientcompany/ClientCompanyReader.java
@@ -1,0 +1,7 @@
+package timf.voc.task.domain.clientcompany;
+
+import java.util.Optional;
+
+public interface ClientCompanyReader {
+	Optional<ClientCompany> get(Long id);
+}
