+++ src/main/java/timf/voc/task/domain/clientcompany/ClientCompanyService.java
@@ -1,0 +1,5 @@
+package timf.voc.task.domain.clientcompany;
+
+public interface ClientCompanyService {
+	ClientCompanyInfo retrieveClientCompany(Long id);
+}
