+++ src/main/java/timf/voc/task/domain/transportcompany/DeliveryDriverPersister.java
@@ -1,0 +1,7 @@
+package timf.voc.task.domain.transportcompany;
+
+import timf.voc.task.domain.transportcompany.aggregate.DeliveryDriver;
+
+public interface DeliveryDriverPersister {
+	void save(DeliveryDriver deliveryDriver);
+}
