+++ src/main/java/timf/voc/task/domain/transportcompany/TransportCompanyInfo.java
@@ -1,0 +1,10 @@
+package timf.voc.task.domain.transportcompany;
+
+import lombok.Data;
+
+@Data
+public class TransportCompanyInfo {
+	private Long id;
+	private String transportCompanyToken;
+	private String companyName;
+}
