+++ src/main/java/timf/voc/task/domain/transportcompany/TransportCompanyPersister.java
@@ -1,0 +1,4 @@
+package timf.voc.task.domain.transportcompany;
+
+public interface TransportCompanyPersister {
+}
