+++ src/main/java/timf/voc/task/domain/transportcompany/TransportCompanyReader.java
@@ -1,0 +1,4 @@
+package timf.voc.task.domain.transportcompany;
+
+public interface TransportCompanyReader {
+}
