+++ src/main/java/timf/voc/task/domain/transportcompany/TransportCompanyService.java
@@ -1,0 +1,4 @@
+package timf.voc.task.domain.transportcompany;
+
+public interface TransportCompanyService {
+}
