+++ src/main/java/timf/voc/task/domain/voc/PenaltyInfo.java
@@ -1,0 +1,7 @@
+package timf.voc.task.domain.voc;
+
+import lombok.Data;
+
+@Data
+public class PenaltyInfo {
+}
