+++ src/main/java/timf/voc/task/domain/voc/PenaltyReader.java
@@ -1,0 +1,4 @@
+package timf.voc.task.domain.voc;
+
+public interface PenaltyReader {
+}
