+++ src/main/java/timf/voc/task/domain/voc/VocReader.java
@@ -1,0 +1,12 @@
+package timf.voc.task.domain.voc;
+
+import java.util.List;
+import java.util.Optional;
+
+import timf.voc.task.domain.voc.aggregate.Voc;
+
+public interface VocReader {
+	List<Voc> get();
+
+	Optional<Voc> get(Long vocId);
+}
