--- src/main/java/timf/voc/task/entity/auditEntity/BaseEntity.java
@@ -1,33 +1,0 @@
-package timf.voc.task.entity.auditEntity;
-
-import java.io.Serializable;
-import java.time.LocalDateTime;
-
-import javax.persistence.Column;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-import org.springframework.format.annotation.DateTimeFormat;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public abstract class BaseEntity implements Serializable {
-
-	@DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
-	@Column(nullable = false, updatable = false)
-	@CreatedDate
-	private LocalDateTime createdAt;
-
-	@LastModifiedDate
-	private LocalDateTime updatedAt;
-
-	private Boolean isDeleted = false;
-}
