--- src/main/java/timf/voc/task/entity/voc/aggregate/ClaimResponsibility.java
@@ -1,14 +1,0 @@
-package timf.voc.task.entity.voc.aggregate;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum ClaimResponsibility {
-	TRANSPORT_COMPANY("운송사"),
-	CLIENT_COMPANY("고객사");
-
-	private final String description;
-
-}
