+++ src/main/java/timf/voc/task/interfaces/claim/ClaimDto.java
@@ -1,0 +1,28 @@
+package timf.voc.task.interfaces.claim;
+
+import lombok.Data;
+import timf.voc.task.config.converter.EntityAndDtoConverter;
+import timf.voc.task.domain.claim.ClaimInfo;
+
+public class ClaimDto {
+
+	@Data
+	public static class ClaimRequest {
+		private String personName;
+		private String content;
+	}
+
+	@Data
+	public static class ClaimResponse {
+		private Long id;
+		private String personName;
+		private String content;
+		private boolean handled;
+
+		public static ClaimResponse from(ClaimInfo claimInfo) {
+			return EntityAndDtoConverter.convert(claimInfo, ClaimResponse.class);
+		}
+	}
+
+
+}
