--- src/main/java/timf/voc/task/service/KafkaProducerService.java
@@ -1,17 +1,0 @@
-package timf.voc.task.service;
-
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.stereotype.Service;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class KafkaProducerService {
-
-	private final KafkaTemplate<String, String> kafkaTemplate;
-
-	public void notifyNewVoc() {
-		kafkaTemplate.send("new-voc", "vocCreated");
-	}
-}
